+++ internal/iac_catalog/blueprint_config/blueprint_config_models.go
@@ -15,22 +15,25 @@ type BluePrintConfig struct {
 	Children          map[string]BluePrintConfig  `json:"children"`
 }
 
+type VariableContent struct {
+	Value          string     `json:"value" faker:"word"`
+	RequiredValues string     `json:"requiredValues" faker:"slice_len=2"`
+	DisplayName    string     `json:"displayName" faker:"word"`
+	HelperText     string     `json:"helperText" faker:"word"`
+	FormConfig     FormConfig `json:"formConfig"`
+}
 type OverrideVariable struct {
-	VariableName   string              `json:"variableName" faker:"word"`
-	Value          string              `json:"value" faker:"word"`
-	RequiredValues string              `json:"requiredValues" faker:"slice_len=2"`
-	DisplayName    string              `json:"displayName" faker:"word"`
-	HelperText     string              `json:"helperText" faker:"word"`
-	FormConfig     FormConfig          `json:"formConfig"`
-	Conditionals   []ConditionalConfig `json:"conditionals"`
+	VariableName string `json:"variableName" faker:"word"`
+	VariableContent
+	Conditionals []ConditionalConfig `json:"conditionals"`
+	IsHidden     bool                `json:"isHidden"` // based on omit variables
+	UsedInHCL    bool                `json:"usedInHCL"`
 }
 
 type ConditionalConfig struct {
-	Source         string `json:"source" faker:"word"`
-	Condition      string `json:"condition" faker:"word"`
-	Type           string `json:"type" faker:"word"`
-	Value          string `json:"value" faker:"word"`
-	RequiredValues string `json:"requiredValues" `
+	Source    string `json:"source" faker:"word"`
+	Condition string `json:"condition" faker:"word"`
+	VariableContent
 }
 
 type FormConfig struct {